Convert sample app to show latest recommended implementation
