--- app/src/main/java/com/pavilionpay/igaming/remote/HttpRoutes.kt
@@ -1,7 +1,0 @@
-package com.pavilionpay.igaming.remote
-
-import com.pavilionpay.igaming.BuildConfig
-
-object HttpRoutes {
-    const val INITIALIZE_PATRON_SESSION = "${BuildConfig.BASE_URL}/api/patronsession/"
-}
