--- app/src/main/java/com/pavilionpay/igaming/remote/PatronResponseDto.kt
@@ -1,6 +1,0 @@
-package com.pavilionpay.igaming.remote
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PatronResponseDto(val sessionId: String)
