--- app/src/main/java/com/pavilionpay/igaming/ui/theme/Color.kt
@@ -1,61 +1,0 @@
-package com.pavilionpay.igaming.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val PrimaryLight = Color(0xff00639b)
-val OnPrimaryLight = Color(0xffffffff)
-val PrimaryContainerLight = Color(0xffcee5ff)
-val OnPrimaryContainerLight = Color(0xff001d33)
-
-val SecondaryLight = Color(0xff51606f)
-val OnSecondaryLight = Color(0xffffffff)
-val SecondaryContainerLight = Color(0xffd5e4f7)
-val OnSecondaryContainerLight = Color(0xff0e1d2a)
-
-val TertiaryLight = Color(0xff68587a)
-val OnTertiaryLight = Color(0xffffffff)
-val TertiaryContainerLight = Color(0xffeedbff)
-val OnTertiaryContainerLight = Color(0xff231533)
-
-val ErrorLight = Color(0xffba1a1a)
-val OnErrorLight = Color(0xffffffff)
-val ErrorContainerLight = Color(0xffffdad6)
-val OnErrorContainerLight = Color(0xff410002)
-
-val BackgroundLight = Color(0xfffcfcff)
-val OnBackgroundLight = Color(0xff1a1c1e)
-val SurfaceLight = Color(0xfffcfcff)
-val OnSurfaceLight = Color(0xff1a1c1e)
-
-val OutlineLight = Color(0xff72777f)
-val SurfaceVariantLight = Color(0xffdee3eb)
-val OnSurfaceVariantLight = Color(0xff42474e)
-
-val PrimaryDark = Color(0xff96ccff)
-val OnPrimaryDark = Color(0xff003353)
-val PrimaryContainerDark = Color(0xff004a76)
-val OnPrimaryContainerDark = Color(0xffcee5ff)
-
-val SecondaryDark = Color(0xffb9c8da)
-val OnSecondaryDark = Color(0xff233240)
-val SecondaryContainerDark = Color(0xff3a4857)
-val OnSecondaryContainerDark = Color(0xffd5e4f7)
-
-val TertiaryDark = Color(0xffd3bfe6)
-val OnTertiaryDark = Color(0xff382a49)
-val TertiaryContainerDark = Color(0xff504061)
-val OnTertiaryContainerDark = Color(0xffeedbff)
-
-val ErrorDark = Color(0xffffb4ab)
-val OnErrorDark = Color(0xff690005)
-val ErrorContainerDark = Color(0xff93000a)
-val OnErrorContainerDark = Color(0xffffdad6)
-
-val BackgroundDark = Color(0xff1a1c1e)
-val OnBackgroundDark = Color(0xffe2e2e5)
-val SurfaceDark = Color(0xff1a1c1e)
-val OnSurfaceDark = Color(0xffe2e2e5)
-
-val OutlineDark = Color(0xff8c9198)
-val SurfaceVariantDark = Color(0xff42474e)
-val OnSurfaceVariantDark = Color(0xffc2c7cf)
