+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">iGaming</string>
+    <string name="app_name">VIP Connect Sample App</string>
 </resources>
