+++ settings.gradle.kts
@@ -11,7 +11,7 @@ pluginManagement {
     plugins {
         val kotlinVersion = extra["kotlin.version"] as String
         kotlin("plugin.serialization").version(kotlinVersion)
-        id("com.android.library") version "8.2.0"
+        id("com.android.library") version "8.5.0"
         id("org.jetbrains.kotlin.android") version "1.9.0"
     }
 }
